+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adjusts in proxy
 * Adjusts in start session for Typebot
+* Added mimetype field when sending media
 
 # 1.5.4 (2023-10-09 20:43)